nixos/syncoid: don't delegate permissions to source's parent
The source dataset may be an entire pool, which doesn't have a parent. In
addition, the source dataset must always exist, so there is no reason to
worry about delegating to its parent. On the other hand, the target dataset
must not be the pool root, so it will always have a parent.