generic-extlinux-compatible: add copyKernels option
Add an option to disable copying kernels, initrds and device trees to the
target directory (/boot). On most systems these days, extlinux.conf is on the
root partition and there is no need to copy files out of the store. With this
patch, the files are still copied by default to maintain backwards
compatibility.