nsjail: run newuidmap/newgidmap from PATH
By default, nsjail tries to run /usr/bin/new{u,g}idmap. This will likely work
on non-NixOS, but will obviously fail on NixOS. Therefore, b38dd07 patched the
package to get these binaries from the shadow package. Unfortunately, this is
not the correct solution because newuidmap and newgidmap need to be setuid
root to work correctly. Currently, uid/gid mapping does not work on any OS.

This patch makes nsjail look for the binaries on PATH. This requires switching
to execve() to execvpe() and specifying the binary names without an absolute
path.