python3Packages: don't propagate build-platform cffi
cffi was added to propagatedNativeBuildInputs to fix cross-compilation
in #142301, but this makes build-platform cffi part of the runtime closure
of all downstream packages. For this reason, propagatedNativeBuildInputs
should almost never be used.

On the other hand, it is sufficient to only add cffi to nativeBuildInputs
to fix cross-compilation. This patch implements this change and removes a
few other propagatedNativeBuildInputs that were unnecessarily added
in #142301.