or-tools: make Python support optional
or-tools is currently broken because 9.4 is not compatible with the latest
version of re2. It is difficult to upgrade to the latest version (9.7) because
it depends on pybind11_protobuf, which has a mostly non-functional CMake build
system (no support for installation). I am working on get pybind11_protobuf
working, but as a stopgap, make it possible to disable Python support. It
doesn't hurt to have this anyway as an option to reduce closure size.